Crate libracerd

libracerd provides an http server and set of completion engines for consumption by rust developer tools. The http server itself is a few JSON endpoints providing completion, definition look-up, and compilation. The endpoints are backed by an object implementing SemanticEngine

Documentation for the HTTP endpoints can be found in the http module header.

This project's source code is available on GitHub.

Modules

engine
Engines provide source analysis for rust code

http
http server for semantic engines

util
Som misc shared functionality

Structs

Config
Configuration flags and values