Module libracerd::engine

Engines provide source analysis for rust code

Reexports

pub use self::racer::Racer;

Modules

racer
SemanticEngine implementation for the racer library

Structs

Buffer

Completion
A possible completion for a location

Context
Context for a given operation.

CursorPosition
Position of the cursor in a text file

Definition
Source file and type information for a found definition

Enums

Error
Error type for semantic engine module

Traits

SemanticEngine
Provide completions, definitions, and analysis of rust source code

Type Definitions

Result
Result type for semantic engine module