Struct libracerd::engine::Context

pub struct Context {
    pub buffers: Vec<Buffer>,
    pub query_cursor: CursorPosition,
    pub query_file: String,
}

Context for a given operation.

All operations require a buffer holding the contents of a file, the file's absolute path, and a cursor position to fully specify the request. This object holds all of those items.

Fields

buffers

query_cursor

query_file

Methods

impl Context

fn new<T>(buffers: Vec<Buffer>, position: CursorPosition, file_path: T) -> Context where T: Into<String>

fn query_path<'a>(&'a self) -> &'a Path