Trait libracerd::engine::SemanticEngine

pub trait SemanticEngine: Send + Sync {
    fn initialize(&self, config: &Config) -> Result<()>;
    fn find_definition(&self, context: &Context) -> Result<Option<Definition>>;
    fn list_completions(&self, context: &Context) -> Result<Option<Vec<Completion>>>;
}

Provide completions, definitions, and analysis of rust source code

Required Methods

fn initialize(&self, config: &Config) -> Result<()>

Perform any necessary initialization.

Only needs to be called once when an engine is created.

fn find_definition(&self, context: &Context) -> Result<Option<Definition>>

Find the definition for the item under the cursor

fn list_completions(&self, context: &Context) -> Result<Option<Vec<Completion>>>

Get a list of completions for the item under the cursor