Module libracerd::http

http server for semantic engines

Structs

EngineProvider
Iron middleware responsible for attaching a semantic engine to each request

Server
Wrapper type with information and control of the underlying HTTP server

Enums

Error
Errors occurring in the http module

Functions

serve
Start the http server using the given configuration

Type Definitions

Result