Function libracerd::http::definition::find

pub fn find(req: &mut Request) -> IronResult<Response>

Given a location, return where the identifier is defined

Possible responses include

200 OK the request was successful and a JSON object is returned.

204 No Content the request was successful, but no match was found.

400 Bad Request the request payload was malformed

500 Internal Server Error some unexpected error occurred